cmake: Permit disabling external http-parser
When attempting to build libgit2 as an isolated static lib, CMake
gleefully attempts to use the system http-parser.  This is typically
seen on Linux systems which install header files with every package,
such as Gentoo.

Allow developers to forcibly disable using the system http-parser with
the config switch USE_EXT_HTTP_PARSER.  Defaults to ON to maintain
previous behavior.

Signed-off-by: Jason Cooper <jason@lakedaemon.net>